Switch to 0.8.8 and drop 0.7.9 support
