+++ package.json
@@ -2,7 +2,7 @@
     "name": "wor",
     "version": "0.1.5",
     "devDependencies": {
-        "@league-of-foundry-developers/foundry-vtt-types": "^0.7.9-6",
+        "@league-of-foundry-developers/foundry-vtt-types": "^0.8.8-9",
         "@rollup/plugin-replace": "^2.4.2",
         "@rollup/plugin-typescript": "^8.2.1",
         "@types/chai": "^4.2.18",
