+++ src/active-effects/disable-expired-effects.ts
@@ -10,7 +10,7 @@ import { wasExpiryTriggeredFor } from './expiry-messages'
  */
 Hooks.on('init', function()
 {
-    CONFIG.ActiveEffect.entityClass = class extends CONFIG.ActiveEffect.entityClass
+    CONFIG.ActiveEffect.documentClass = class extends CONFIG.ActiveEffect.documentClass
     {
         override get isTemporary(): boolean
         {
