+++ src/active-effects/index.ts
@@ -2,4 +2,3 @@ import './check-for-expiry'
 import './disable-expired-effects'
 import './expiry-messages'
 import './prettier-durations'
-import './record-start-time'
